finai-hackops/backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
passlib[bcrypt]
python-multipart
python-jose[cryptography]
pydantic[email]
